🗺️ Atlas: Splitting the Query Executor Iterators Blob
🕸️ Tangle: `src/query/executor/iterators.rs` is a 5500-line "Blob" module containing both core iteration logic and over 2500 lines of tests.

📐 Blueprint: Refactored into `src/query/executor/iterators/` directory. Moved tests to `tests.rs`, leaving the core logic in `mod.rs`.

🧱 Stability: Reduced file size and cognitive load, strict separation enforced.

🔬 Verification: Builds successfully, tests pass.
